WIP: adding basic tests
